flatten modules (again rust 2018?)
